refactor initialized wallets interface
