+++ packages/core/standard/src/window.ts
@@ -1,62 +1,24 @@
 import type { Wallet } from './wallet.js';
 
-/** Global `window` containing a `navigator.wallets` object. */
+/** Global `window` containing a `navigator.wallets` interface. */
 export interface WalletsWindow extends Window {
     navigator: WalletsNavigator;
 }
 
-/** Global `window.navigator` containing a `wallets` object. */
+/** Global `window.navigator` containing a `wallets` interface. */
 export interface WalletsNavigator extends Navigator {
-    wallets?: NavigatorWallets;
+    wallets?: Wallets;
 }
 
-/** Global `window.navigator.wallets` object or callback array. */
-export type NavigatorWallets = WalletsCallback[] | Wallets;
-
-/**
- * TODO: docs
- */
-export type WalletsCallback = (wallets: Pick<Wallets, 'register'>) => void;
-
-/** Global `window.navigator.wallets` object API. */
+/** Global `window.navigator.wallets` interface. */
 export interface Wallets {
     /**
      * TODO: docs
      */
     push(...callbacks: ReadonlyArray<WalletsCallback>): void;
-
-    /**
-     * TODO: docs
-     */
-    register(...wallets: ReadonlyArray<Wallet>): () => void;
-
-    /**
-     * TODO: docs
-     */
-    get(): ReadonlyArray<Wallet>;
-
-    /**
-     * TODO: docs
-     */
-    on<E extends WalletsEventNames = WalletsEventNames>(event: E, listener: WalletsEvents[E]): () => void;
 }
 
-/** Events emitted by the global `wallets` object. */
-export interface WalletsEvents {
-    /**
-     * Emitted when wallets are registered.
-     *
-     * @param wallets Wallets that were registered.
-     */
-    register(...wallets: ReadonlyArray<Wallet>): void;
-
-    /**
-     * Emitted when wallets are unregistered.
-     *
-     * @param wallets Wallets that were unregistered.
-     */
-    unregister(...wallets: ReadonlyArray<Wallet>): void;
-}
-
-/** TODO: docs */
-export type WalletsEventNames = keyof WalletsEvents;
+/**
+ * TODO: docs
+ */
+export type WalletsCallback = (wallets: { register(...wallets: ReadonlyArray<Wallet>): () => void }) => void;
